feat(auth): complete user authentication flow
- Add email verification
- Fix sign-in routing issues
- Update middleware configuration
- Configure clerk authentication
- Add proper redirects
