+++ src/app/auth/sign-in/sso-callback/google/page.tsx
@@ -1,27 +1,15 @@
 'use client';
 
-import { useClerk } from '@clerk/nextjs';
-import { useEffect } from 'react';
+import { AuthenticateWithRedirectCallback } from '@clerk/nextjs';
 
 export default function GoogleSSOCallback() {
-  const { handleRedirectCallback } = useClerk();
-
-  useEffect(() => {
-    async function handleCallback() {
-      try {
-        await handleRedirectCallback();
-        window.location.href = '/';
-      } catch (error) {
-        console.error('Google callback error:', error);
-        window.location.href = '/auth/sign-in';
-      }
-    }
-    handleCallback();
-  }, []); // 移除 handleRedirectCallback 依赖
-
   return (
     <div className="min-h-screen flex items-center justify-center">
       <div className="animate-pulse">正在完成 Google 登录...</div>
+      <AuthenticateWithRedirectCallback
+        afterSignInUrl="/dashboard"
+        afterSignUpUrl="/dashboard"
+      />
     </div>
   );
 }
