+++ src/app/auth/sign-up/sso-callback/google/page.tsx
@@ -2,13 +2,13 @@
 
 import { AuthenticateWithRedirectCallback } from '@clerk/nextjs';
 
-export default function SSOCallback() {
+export default function GoogleSignUpSSOCallback() {
   return (
     <div className="min-h-screen flex items-center justify-center">
+      <div className="animate-pulse">正在完成 Google 注册...</div>
       <AuthenticateWithRedirectCallback
         afterSignInUrl="/dashboard"
         afterSignUpUrl="/dashboard"
-        redirectUrl={`${window.location.origin}/dashboard`}
       />
     </div>
   );
