+++ src/app/auth/sign-up/sso-callback/page.tsx
@@ -2,13 +2,13 @@
 
 import { AuthenticateWithRedirectCallback } from '@clerk/nextjs';
 
-export default function SSOCallback() {
+export default function SignUpSSOCallback() {
   return (
     <div className="min-h-screen flex items-center justify-center">
+      <div className="animate-pulse">处理注册中...</div>
       <AuthenticateWithRedirectCallback
         afterSignInUrl="/dashboard"
         afterSignUpUrl="/dashboard"
-        redirectUrl={`${window.location.origin}/dashboard`}
       />
     </div>
   );
