--- src/app/contexts/ToastContext.js
@@ -1,32 +1,0 @@
-'use client';
-
-import React, { createContext, useContext, useState } from 'react';
-import Toast from '../../components/ui/Toast'; // 注意这里的路径修改
-
-export const ToastContext = createContext({
-  showToast: () => {},
-});
-
-export function ToastProvider({ children }) {
-  const [toast, setToast] = useState(null);
-
-  const showToast = (message, type = 'info') => {
-    setToast({ message, type });
-    setTimeout(() => setToast(null), 3000);
-  };
-
-  return (
-    <ToastContext.Provider value={{ showToast }}>
-      {children}
-      {toast && (
-        <Toast
-          message={toast.message}
-          type={toast.type}
-          onClose={() => setToast(null)}
-        />
-      )}
-    </ToastContext.Provider>
-  );
-}
-
-export const useToast = () => useContext(ToastContext);
